requirements.txt
asgiref==3.2.7
astroid==2.4.2
autopep8==1.5.3
certifi==2020.4.5.2
chardet==3.0.4
colorama==0.4.3
Django==3.0.7
django-cors-headers==3.3.0
django-filter==2.3.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
flake8==3.8.3
idna==2.9
isort==4.3.21
jedi==0.17.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
parso==0.7.0
pluggy==0.13.1
pycodestyle==2.6.0
pydocstyle==5.0.2
pyflakes==2.2.0
PyJWT==1.7.1
pylint==2.5.3
python-jsonrpc-server==0.3.4
python-language-server==0.33.1
pytz==2020.1
requests==2.23.0
rope==0.17.0
six==1.15.0
snowballstemmer==2.0.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.9
wrapt==1.12.1
yapf==0.30.0